Add files via upload
Objective - To develop a MATLAB program that help us visualize shear stress distribution in a rectangular beam.

Theory - It is assumed that when a beam is subjected to force uniform shear stresses is developed, but the fact is shear stress is distributed parabolically, maximum at neutral fiber. 

Result - From the figure we can see that shear stress is varying parabolically, maximum value at middle at neutral axis.